Fix msvc int conversion warnings
VHAT-1813

Following the google code style requirements to integer types:
- use 'uint64_t' for message id, because it should be big and have defined
  2^64 overflow.
- use 'int' for timeouts, as we don't expect timeouts of more then 2^31
- use 'int' for all coordinates

+++ ATE_common/include/message_factory/json_utils.h
@@ -104,8 +104,8 @@ void ExtractTapObjectRequestParams(const Json::Value& params, int& x, int& y, sq
  * @param error - error object, null on successs
  **/
 
-void ExtractLongPressRequestParams(const Json::Value& params, uint16_t& x, uint16_t& y,
-                                   std::chrono::milliseconds& timeout, Json::Value& error);
+void ExtractLongPressRequestParams(const Json::Value& params, int& x, int& y, std::chrono::milliseconds& timeout,
+                                   Json::Value& error);
 
 /**
  * @brief Extract params for 'TouchAndDrag' method
