+++ ATE_common/src/message_factory/json_utils.cpp
@@ -174,8 +174,8 @@ void ExtractTapObjectRequestParams(const Json::Value& params, int& x, int& y, sq
   }
 }
 
-void ExtractLongPressRequestParams(const Json::Value& params, uint16_t& x, uint16_t& y,
-                                   std::chrono::milliseconds& timeout, Json::Value& error) {
+void ExtractLongPressRequestParams(const Json::Value& params, int& x, int& y, std::chrono::milliseconds& timeout,
+                                   Json::Value& error) {
   error = Json::Value{};
 
   try {
